[Job Scheduler] Initial commit with draft skeleton and Makefile
Start building the skeleton for the job scheduler project.
Begin with a .cpp file for each tool/program in the system.
A setup.sh script is also provided to bootstrap a virtual machine
with the requisite dependencies to build the project. An initial
makefile to build the programs is also provided for students.
Later commits will start filling out the skeleton with more details
to be filled in for the project implementation.